Extract Memes tab button to keep page within size limit
